requirements.txt
pytest
selenium
click
pyyaml
pytest-xdist
requests
allure-pytest